Improving "reset" and "add matrix"
Reset button is not working properly
Now using new_matr_alt.js as the js file. Different way to create new fields that seems better. Also adds a "Remove" button after each group of new fields (not working at the moment).
"count" variable changes with the use of each button - with "reset", count goes back to 1; with "Remove" count = count -1 (if with each new group of fields count +=1, then if we remove a group we have to do count -=1 )