Saves doc according to the matrices selected
The user can have input for several matrices but only the "checked" ones will be saved to the final document.
Created an array with a row for each group of input (the first 6 fields + each 3 fields representing a new matrix). 
Each group has a value of 1 or 0. If 1, the checkbox is checked, if 0, it is unchecked.
Only the groups with value 1 (checked) will be saved to the doc