Added matrix creator + some small changes
Added "number of demes" field
"Time of change" for the first matrix is always zero
Changed order of some fields
Added "create matrix" button. This will create a matrix where the number of rows and columns is the number of demes. (hitting enter after writing the number of demes also creates the matrix)
Working on making the matrix editable

Not saving