Saving files
The program is working. It saves the input to a file.
If the name of the file already exists, the program does not save and the user should choose another name for the file. (in the future it will probably show a message telling the user to change the name)
Shows a message if the first two fields are not filled and "ok" button is clicked
Will remove the space at the end of some lines
Will remove extra line at the end of the file